hanbon_python_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai
httpx==0.25.0
mcp
pydantic==2.5.0
asyncio-mqtt==0.13.0
beautifulsoup4==4.12.2
Pillow==10.1.0
numpy==1.24.3
redis==5.0.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
uvicorn==0.24.0
fastapi
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
markdown==3.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3 